Move Media3 player sample to demo app
