+++ beforeafter/build.gradle
@@ -35,11 +35,7 @@ afterEvaluate {
 dependencies {
     implementation libs.compose.extended.gestures
 
-    implementation libs.androidx.compose.ui.ui.tooling
     implementation libs.androidx.runtime
-
-    compileOnly libs.androidx.media3.exoplayer
-    compileOnly libs.androidx.media3.ui.compose
 }
